DrawShowing: mRenderedStates is compared unsigned (94.5 -> 94.6%)
The target's guard is 'cmplwi cr6, r11, 0x1'; ours was 'cmpwi'.  Casting at
the comparison gets the opcode.

Tried and reverted first: changing the member to 'unsigned int'.  That is the
more natural reading but it regressed 92.5% -> 92.4% at the time, because the
'mRenderedStates |= state' sites all changed shape too.  The cast is narrow
and does not touch them.

+++ src/system/rndobj/TexBlender.cpp
@@ -201,7 +201,7 @@ void RndTexBlender::DrawShowing() {
                 }
             }
             if (unkc0 || !nearList.empty() || !farList.empty() || !customList.empty()
-                || mRenderedStates != 1) {
+                || (unsigned int)mRenderedStates != 1) {
                 unkc0 = false;
                 RndCam *cam = TheRnd.GetDefaultCam();
                 RndCam *prevCam = RndCam::Current();
